+++ wingspanscorecard/src/app/ui/add-game.tsx
@@ -2,7 +2,7 @@
 
 import { useState } from 'react';
 import { Player } from '@/app/lib/definitions';
-import { createGame } from '@/app/lib/action';
+import { createGame } from '@/app/lib/server-uploads';
 import ScoreColumn from '@/app/ui/score-column';
 
 export default function Form({ players }: { players: Player[] }) {
